✅ PIC-1194: Disable auth for pact tests
