+++ src/main/kotlin/uk/gov/justice/digital/hmpps/hmppsuserpreferences/config/ResourceServerConfiguration.kt
@@ -1,21 +1,17 @@
 package uk.gov.justice.digital.hmpps.hmppsuserpreferences.config
 
-import com.nimbusds.jwt.JWTParser
-import org.springframework.context.annotation.Bean
 import org.springframework.context.annotation.Configuration
 import org.springframework.context.annotation.Profile
 import org.springframework.security.config.annotation.web.builders.HttpSecurity
 import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity
 import org.springframework.security.config.annotation.web.configuration.WebSecurityConfigurerAdapter
 import org.springframework.security.config.http.SessionCreationPolicy
-import org.springframework.security.oauth2.jwt.Jwt
-import org.springframework.security.oauth2.jwt.JwtDecoder
-import org.springframework.security.oauth2.jwt.MappedJwtClaimSetConverter
 import org.springframework.security.oauth2.server.resource.authentication.JwtAuthenticationConverter
 import org.springframework.security.oauth2.server.resource.authentication.JwtGrantedAuthoritiesConverter
 
 @Configuration
 @EnableWebSecurity
+@Profile("!unsecured")
 class ResourceServerConfiguration : WebSecurityConfigurerAdapter() {
   override fun configure(http: HttpSecurity) {
     http
@@ -39,7 +35,6 @@ class ResourceServerConfiguration : WebSecurityConfigurerAdapter() {
       .oauth2ResourceServer().jwt().jwtAuthenticationConverter(jwtAuthenticationConverter())
   }
 
-  @Bean
   fun jwtAuthenticationConverter(): JwtAuthenticationConverter {
     // hmpps auth tokens have roles in a custom `authorities` claim.
     // the authorities are already prefixed with `ROLE_`.
@@ -50,27 +45,5 @@ class ResourceServerConfiguration : WebSecurityConfigurerAdapter() {
     val jwtAuthenticationConverter = JwtAuthenticationConverter()
     jwtAuthenticationConverter.setJwtGrantedAuthoritiesConverter(grantedAuthoritiesConverter)
     return jwtAuthenticationConverter
-  }
-
-  @Bean
-  @Profile("test")
-  fun testJwtDecoder(): JwtDecoder {
-    return TestJwtDecoder()
-  }
-}
-
-internal class TestJwtDecoder : JwtDecoder {
-  private val claimSetConverter = MappedJwtClaimSetConverter.withDefaults(emptyMap())
-
-  override fun decode(token: String): Jwt? {
-    // extract headers and claims, but do not attempt to verify signature
-    val jwt = JWTParser.parse(token)
-    val headers = LinkedHashMap<String, Any>(jwt.header.toJSONObject())
-    val claims = claimSetConverter.convert(jwt.getJWTClaimsSet().claims)
-
-    return Jwt.withTokenValue(token)
-      .headers { it.putAll(headers) }
-      .claims { it.putAll(claims) }
-      .build()
   }
 }
